Server starts and responds, still errors
The server starts and responds, but assets still need to be precompiled.

It seems like it wasn't working before because I was trying to run
`bundle install` during the Dockerfile build process, but I hadn't
copied in any of the files yet. So running it in the
`deploy-production.sh` script seems to work